Move findPncCourtCase function into root lib folder
