+++ packages/core/lib/findPncCourtCase.test.ts
@@ -1,4 +1,4 @@
-import generateAhoMatchingPncAdjudicationAndDisposals from "../tests/helpers/generateAhoMatchingPncAdjudicationAndDisposals"
+import generateAhoMatchingPncAdjudicationAndDisposals from "../phase2/tests/helpers/generateAhoMatchingPncAdjudicationAndDisposals"
 import findPncCourtCase from "./findPncCourtCase"
 
 describe("findPncCourtCase", () => {
