+++ packages/core/phase2/lib/areAnyPncDisposalsWithType.ts
@@ -1,6 +1,6 @@
 import type { AnnotatedHearingOutcome, Offence } from "../../types/AnnotatedHearingOutcome"
 
-import findPncCourtCase from "./findPncCourtCase"
+import findPncCourtCase from "../../lib/findPncCourtCase"
 
 const areAnyPncDisposalsWithType = (aho: AnnotatedHearingOutcome, offence: Offence, disposalType: number): boolean => {
   const matchingPncCourtCase = findPncCourtCase(aho, offence)
