+++ packages/core/phase3/lib/hearingDetails/getConvictionDateFromPncAdjudicationIfOffenceIsAdjournedSineDie.ts
@@ -1,6 +1,6 @@
 import type { AnnotatedHearingOutcome, Offence } from "../../../types/AnnotatedHearingOutcome"
 
-import findPncCourtCase from "../../../phase2/lib/findPncCourtCase"
+import findPncCourtCase from "../../../lib/findPncCourtCase"
 
 const ADJOURNED_SINE_DIE_DISPOSAL_CODE = 2007
